refactor: use custom build output
