+++ src/api/_lib/handlers/tagOrBanrch.ts
@@ -1,7 +1,7 @@
 import type { CreateEvent, DeleteEvent } from '@octokit/webhooks-types';
-import { PackageName } from '../utils/constants';
-import { getPotentialPackageTarget } from '../utils/functions';
-import type { DiscordWebhooksTarget } from '../utils/webhooks';
+import { PackageName } from '../utils/constants.js';
+import { getPotentialPackageTarget } from '../utils/functions.js';
+import type { DiscordWebhooksTarget } from '../utils/webhooks.js';
 
 /**
  * Gets the target for incoming create or delete type webhooks
