+++ tsconfig.eslint.json
@@ -13,5 +13,5 @@
 		"**/*.spec.js",
 		"**/*.spec.mjs"
 	],
-	"exclude": ["middleware.ts"]
+	"exclude": ["src/middleware.ts"]
 }
